requirements.txt
Django==4.2.19
minio==7.2.13
qrcode==8.0
pillow==11.0.0
pytz==2024.2
gunicorn==23.0.0
psycopg==3.2.4
# psycopg2==2.9.10
psycopg2-binary==2.9.10
loguru==0.7.3
openpyxl==3.1.2